requirements.txt
flask
openai==1.50.2
httpx==0.28.1
anthropic>=0.26.0
google-genai
python-dotenv
requests
beautifulsoup4
pydantic
gunicorn